This PR addresses more review comments from the CMSSW PR. These are the things I have changed. I'll update this comment if I add more commits.

- Atomic operations:
  - https://github.com/cms-sw/cmssw/pull/45117#discussion_r1681629998
  - https://github.com/cms-sw/cmssw/pull/45117#discussion_r1681674098
  - https://github.com/cms-sw/cmssw/pull/45117#discussion_r1681692689
  - https://github.com/cms-sw/cmssw/pull/45117#discussion_r1681705133
  - https://github.com/cms-sw/cmssw/pull/45117#discussion_r1681722349
  
  I changed the syntax to the suggested one and specified `hierarchy::Threads` explicitly. The plots look the same and the timing might be marginally better, so it seems fine.

- Map lookups:
  - https://github.com/cms-sw/cmssw/pull/45117#discussion_r1681636712

  I refactored it so that it only performs one lookup.

- ROCm support:
  - https://github.com/cms-sw/cmssw/pull/45117#discussion_r1682937131

  I added an extra include so that the fp16 header is also imported for ROCm.


Here's a timing comparison.
```
This PR
Average time for map loading = 501.109 ms
Average time for input loading = 7555.69 ms
Average time for lst::Event creation = 0.00690278 ms
   Evt    Hits       MD       LS      T3       T5       pLS       pT5      pT3      TC       Reset    Event     Short             Rate
   avg     17.8      1.3      0.4      1.4      1.5      0.4      0.8      0.6      1.2      0.1      25.5       7.3+/-  1.4      27.6   explicit_cache[s=1]
   avg      5.7      1.4      0.5      1.9      1.9      0.4      1.3      0.9      1.7      0.2      15.9       9.7+/-  2.0       9.2   explicit_cache[s=2]
   avg     14.1      1.8      0.8      3.0      2.8      0.5      2.3      1.3      2.6      0.3      29.7      15.0+/-  3.2       8.1   explicit_cache[s=4]
   avg     17.1      2.3      1.1      4.1      4.3      0.6      3.4      2.0      3.7      0.6      39.2      21.5+/-  6.5       7.1   explicit_cache[s=6]
   avg     21.2      3.0      1.6      5.7      5.9      0.9      5.0      2.8      4.7      0.8      51.4      29.3+/-  7.3       6.9   explicit_cache[s=8]

Current batch5 branch (d3549cba529)
Average time for map loading = 497.301 ms
Average time for input loading = 7684.05 ms
Average time for lst::Event creation = 0.00804947 ms
   Evt    Hits       MD       LS      T3       T5       pLS       pT5      pT3      TC       Reset    Event     Short             Rate
   avg     17.0      1.4      0.4      1.4      1.5      0.4      0.8      0.6      1.2      0.1      24.8       7.4+/-  1.4      26.9   explicit_cache[s=1]
   avg      5.5      1.5      0.5      2.0      1.9      0.4      1.3      0.9      1.6      0.2      15.9       9.9+/-  2.1       9.1   explicit_cache[s=2]
   avg      9.1      1.8      0.8      3.1      3.1      0.5      2.2      1.2      2.6      0.4      24.8      15.1+/-  3.4       6.9   explicit_cache[s=4]
   avg     15.9      2.3      1.1      4.1      4.0      0.6      3.3      2.0      3.7      0.6      37.7      21.2+/-  5.7       6.8   explicit_cache[s=6]
   avg     22.2      2.8      1.3      5.5      5.4      0.8      4.5      2.5      5.0      0.7      50.6      27.7+/-  7.6       6.8   explicit_cache[s=8]
```