@ariostas 2024-08-14
I was thinking of adding another commit to this PR addressing the namespace comments, but it's turning out to be significantly lengthier than I expected and it will probably also require renaming the `Event` class, so I think it's better to do it as its own PR. So I'll mark this PR as ready to review.